Add support to clean up old blobs (finally)
Excited to get this in. Basically it works like the old dup functionality, we now also move any old refs and any unreferenced blobs to a flat directory of dup's/refs/blobs that are keyed by the current date and time. The user can just rm -rf the directory, and that should be great, or they can look through it and make sure nothing is wrong.